requirements.txt
streamlit==1.37.1
pandas==2.2.2
scikit-learn==1.5.1
python-dateutil==2.9.0.post0
